Numerous structural refactorings: 1)Move proxy related code to src/proxy 2)Rename pipeline to request-pipeline, PipelineContext to RequestPipelineContext 3)Move destination request-related code to src/request-pipeline/destination-request 4)Extract encoding/decoding logic from content utils and move to src/processing/encoding. Move Charset there as well.
