+++ src/request-pipeline/destination-request/windows-domain.js
@@ -1,5 +1,5 @@
 import childProcess from 'child_process';
-import promisify from '../utils/promisify';
+import promisify from '../../utils/promisify';
 
 var exec   = promisify(childProcess.exec);
 var cached = null;
