+++ src/request-pipeline/xhr/headers.js
