+++ src/request-pipeline/xhr/same-origin-policy.js
