+++ test/server/windows-domain-test.js
@@ -1,6 +1,6 @@
 var platform      = require('os').platform;
 var expect        = require('chai').expect;
-var windowsDomain = require('../../lib/pipeline/windows-domain');
+var windowsDomain = require('../../lib/request-pipeline/destination-request/windows-domain');
 
 var IS_WINDOWS = /^win/.test(platform());
